Refactor repository unit tests
Deleted old unit test suite and rewrite test_repo.py to be more feature
oriented.

This changeset includes a fix on how repositories label updates are
handled as there was an issue related to the type returned by DiffDict.

The tests for the repo hooks are not complete as that functionality will
probably need a "re-do" given that the "name" of the hooks are used as
key and all webhooks should be named just "web" according to github
documentation. To be reviewed and fixed in a future commit.