Refactor organization unit tests
Re-do of the unit tests that validate the code in organization.py, done
to be more consistent with the way repository is tested and ease the
maintenance.

Tests for teams and hooks are missing/in progress as there is an
issue on the way it is designed:

- teams: The id is exposed, which prevents creation of teams
- hooks: name is used as key, but all webhooks have "web" as name

Even the work cannot be considered complete, this validates further than
the previous test-suite. Once the mentioned issues are fixed and the
suite is completed, much of the tests in the integration folder can be
removed as this suite covers it better.