Dont expose team id
Exposing the team id is inconsistence with the rest of the elemnts.
The same way the label id or the hook id is not exposed neither should
be the team id.
In addition to consistency it makes cucumbersome to create a new team as
the API does not accept the ID at creation time.